Add UNet model variations with variable number of classes
This commit adds two variations of the UNet model: UNet_Var and VovUnet_Var. These models have an additional classification branch and can handle a variable number of segmentation and classification classes. The models are implemented in the `unet_model.py` file and use the `unet_parts.py` file for the U-Net model components.